move custom_alloc type defs
