+++ src/local_array/store/mod.rs
@@ -1,5 +1,7 @@
 pub mod custom_alloc;
-pub(crate) mod default_store;
 pub mod iterators;
+
+pub(crate) mod default_store;
+pub(crate) mod atomic_types;
 
 pub use default_store::DefaultStore;
