+++ src/prelude/mod.rs
@@ -1,7 +1,7 @@
 pub use crossbeam_epoch::{self as epoch, Guard};
 
-pub use routecore::bgp::PrefixRecord;
 pub use routecore::addr::Prefix;
+pub use routecore::bgp::PrefixRecord;
 pub use routecore::record::{Meta, Record};
 
 pub use crate::{AddressFamily, IPv4, IPv6};
@@ -10,8 +10,9 @@ pub use rotonda_macros::create_store;
 pub use rotonda_macros::stride_sizes;
 
 pub use crate::custom_alloc::CustomAllocStorage;
-pub use crate::custom_alloc::{NodeBuckets, PrefixBuckets};
-pub use crate::custom_alloc::{NodeSet, PrefixSet};
+pub use crate::local_array::store::atomic_types::{
+    NodeBuckets, NodeSet, PrefixBuckets, PrefixSet,
+};
 
 pub use crate::local_array::tree::{PrefixId, StrideNodeId, TreeBitMap};
 pub use crate::stride::{Stride3, Stride4, Stride5};
